Added multiuser

+++ app/routes/tasks.py
@@ -6,22 +6,21 @@ tasks_bp=Blueprint('tasks',__name__)
 
 @tasks_bp.route('/')
 def view_tasks():
-    if 'user' not in session:
+    if 'user_id' not in session:
         return redirect(url_for('auth.login'))
 
-    
-    tasks=Task.query.all()
+    tasks = Task.query.filter_by(user_id=session['user_id']).all()
     return render_template('tasks.html',tasks=tasks)
 
 @tasks_bp.route('/add',methods=["POST"])
 def add_task():
-    if 'user' not in session:
+    if 'user_id' not in session:
         return redirect(url_for('auth.login'))
     
 
     title=request.form.get('title')
     if title:
-        new_task=Task(title=title,status="Pending")
+        new_task=Task(title=title,status="Pending",user_id=session['user_id'])
         db.session.add(new_task)
         db.session.commit()
         flash('Task added successfully','success')
@@ -30,8 +29,10 @@ def add_task():
 
 @tasks_bp.route('/toggle/<int:task_id>',methods=["POST"])
 def toggle_status(task_id):
+    if 'user_id' not in session:
+        return redirect(url_for('auth.login'))
     task=Task.query.get(task_id)
-    if task:
+    if task and task.user_id==session['user_id']:
         if task.status=='Pending':
             task.status = 'Working'
         elif task.status == 'Working':
@@ -39,14 +40,17 @@ def toggle_status(task_id):
         else:
             task.status ='Pending'
 
-        db.session.commit()
-        return redirect(url_for('tasks.view_tasks') + f"#task-{task.id}")
+            db.session.commit()
+    return redirect(url_for('tasks.view_tasks') + f"#task-{task.id}")
 
 
 
 @tasks_bp.route('/clear',methods=["POST"])
 def clear_tasks():
-    Task.query.delete()
+    if 'user_id' not in session:
+        return redirect(url_for('auth.login'))
+    
+    Task.query.filter_by(user_id=session['user_id']).delete()
     db.session.commit()
     flash('All tasks cleared!','info')
     return redirect(url_for('tasks.view_tasks'))
